+++ CppBasic/MultiSensorSystem/Sensors/TemperatureSensor.h
@@ -7,5 +7,5 @@ public:
     TemperatureSensor(SensorType type, const std::string &name, float minRange, float maxRange)
         : Sensor(type, name, minRange, maxRange) {
     }
-    float read() const override;
+    double read() const override;
 };
